ResLogger2.Web/services/web/requirements.txt
Flask==2.1.1
jsonschema==4.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
Flask-SQLAlchemy==2.5.1
sqlalchemy-batch-inserts==0.0.4
pony==0.7.16
tortoise-orm[asyncpg]==0.19.0